Moved setup_residues to tidy and renamed io
