+++ package/CMakeLists.txt
@@ -103,8 +103,8 @@ add_library(
         ${WRK_DIR}/src/cpp/nucleofind/backbone-tracing.h
         ${WRK_DIR}/src/cpp/nucleofind/fragment-library.cpp
         ${WRK_DIR}/src/cpp/nucleofind/fragment-library.h
-        ${WRK_DIR}/src/cpp/nucleofind/io.cpp
-        ${WRK_DIR}/src/cpp/nucleofind/io.h
+        ${WRK_DIR}/src/cpp/nucleofind/input.cpp
+        ${WRK_DIR}/src/cpp/nucleofind/input.h
 
 )
