+++ package/src/python/nautilus-run.h
@@ -20,11 +20,11 @@
 #include "../cpp/nautilus-findml.h"
 #include "../cpp/nucleofind/nucleofind.h"
 #include "../cpp/nucleofind/predicted-maps.h"
-#include "../cpp/nucleofind/io.h"
+#include "../cpp/nucleofind/input.h"
 
-clipper::MiniMol &
-run_cycle(int nhit, double srchst, int verbose, NucleicAcidTargets &natools, const clipper::MMoleculeSequence &seq_wrk,
-          clipper::MiniMol &mol_wrk, const clipper::Xmap<float> &xwrk, NautilusLog &log, bool strong_prune) {
+inline clipper::MiniMol &
+run_cycle(const int verbose, const NucleicAcidTargets &natools, const clipper::MMoleculeSequence &seq_wrk,
+          clipper::MiniMol &mol_wrk, const clipper::Xmap<float> &xwrk, NautilusLog &log, const bool strong_prune) {
     // grow chains
 
     int nas_found = NautilusUtil::count_nas(mol_wrk);
@@ -33,7 +33,6 @@ run_cycle(int nhit, double srchst, int verbose, NucleicAcidTargets &natools, con
     mol_wrk = natools.grow(xwrk, mol_wrk, 25, 0.001);
     log.log("GROW", mol_wrk, verbose >= 5);
 
-    // NautilusUtil::save_minimol(mol_wrk, "grow.pdb");
     ModelTidy::chain_renumber(mol_wrk, seq_wrk);
     NucleicAcidTools::chain_sort(mol_wrk);
     NucleicAcidTools::residue_label(mol_wrk);
@@ -44,12 +43,10 @@ run_cycle(int nhit, double srchst, int verbose, NucleicAcidTargets &natools, con
         mol_wrk = na_join.join(mol_wrk);
         log.log("JOIN", mol_wrk, verbose >= 5);
     }
-//    NautilusUtil::save_minimol(mol_wrk, "join.pdb");
+
     // link
     mol_wrk = natools.link(xwrk, mol_wrk);
     log.log("LINK", mol_wrk, verbose >= 5);
-
-    // NautilusUtil::save_minimol(mol_wrk, "link.pdb");
 
     // prune
     if (strong_prune) {
@@ -58,8 +55,6 @@ run_cycle(int nhit, double srchst, int verbose, NucleicAcidTargets &natools, con
         mol_wrk = natools.prune(mol_wrk);
     }
     log.log("PRUNE", mol_wrk, verbose >= 5);
-
-
 
     // rebuild
     mol_wrk = natools.rebuild_chain(xwrk, mol_wrk);
@@ -74,45 +69,10 @@ run_cycle(int nhit, double srchst, int verbose, NucleicAcidTargets &natools, con
     NucleicAcidRebuildBases na_bases;
     mol_wrk = na_bases.rebuild_bases(xwrk, mol_wrk);
     log.log("BASES", mol_wrk, verbose >= 5);
-    // exit(-1);
 
     return mol_wrk;
 }
 
-clipper::MiniMol setup_residue_types(clipper::Xmap<float> xwrk, clipper::MiniMol mol_wrk) {
-    const clipper::String basetypes = "ACGTU";
-    clipper::MiniMol mol_new(xwrk.spacegroup(), xwrk.cell());
-    for (int c = 0; c < mol_wrk.size(); c++) {
-        clipper::MPolymer mpx = mol_wrk[c];
-        if (!mpx.exists_property("NON-NA")) {
-            bool dna = false;
-            for (int r = 0; r < mpx.size(); r++) {
-                const clipper::String type = mpx[r].type().trim() + " ";
-                const char ctype = type[0];
-                int t = NucleicAcidTools::base_index(ctype);
-                if (t >= 0) mpx[r].set_type("  " + basetypes.substr(t, 1));
-                else mpx[r].set_type("  U");
-                if (mpx[r].type().trim() == "T") dna = true;
-            }
-            // for DNA, prefix type with D and remove O2'
-            if (dna) {
-                for (int r = 0; r < mpx.size(); r++) {
-                    const clipper::String type = " D" + mpx[r].type().trim();
-                    mpx[r].set_type(type);
-                    const int io2 = mpx[r].lookup(" O2'", clipper::MM::ANY);
-                    clipper::MMonomer mm;
-                    mm.copy(mpx[r], clipper::MM::COPY_MP);
-                    for (int a = 0; a < mpx[r].size(); a++) {
-                        if (a != io2) mm.insert(mpx[r][a]);
-                    }
-                    mpx[r] = mm;
-                }
-            }
-        }
-        mol_new.insert(mpx);
-    }
-    return mol_new;
-}
 
 void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycles) {
 
@@ -132,6 +92,7 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
     NucleicAcidTargets natools;
     natools.add_pdb(ippdb_ref);
 
+    // setup
     auto mtz_file = NucleoFind::IO::MTZ(input, res_in);
     auto prediction_files = NucleoFind::IO::Predictions(input);
     auto sequence_file = NucleoFind::IO::Sequence(input);
@@ -151,8 +112,8 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
     mol_wrk = NucleicAcidTools::flag_chains(mol_wrk);
     clipper::MiniMol mol_wrk_original = mol_wrk;
 
+    // find with NucleoFind predictions
     NucleoFind::PredictedMaps predicted_maps = prediction_files.get_predictions();
-
     NucleoFind::Find find = {xwrk, predicted_maps};
     mol_wrk = find.find(mol_wrk);
     log.log("FIND ML", mol_wrk, verbose >= 5);
@@ -171,7 +132,7 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
 
         for (int cyc = 0; cyc < cycles; cyc++) {
             std::cout << "ML Based cycle " << clipper::String(cyc + 1, 3) << std::endl << std::endl;
-            mol_wrk = run_cycle(nhit, srchst, verbose, natools, seq_wrk, mol_wrk, xwrk, log, false);
+            mol_wrk = run_cycle(verbose, natools, seq_wrk, mol_wrk, xwrk, log, false);
             NucleicAcidTools::chain_label(mol_wrk, clipper::MMDBManager::CIF);
             mol_wrk = NucleicAcidTools::chain_sort(mol_wrk);
             NucleicAcidTools::residue_label(mol_wrk);
@@ -179,6 +140,7 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
     }
 
     clipper::MiniMol best_model = mol_wrk;
+    clipper::MiniMol ml_model = mol_wrk;
 
     int best_na_count = NautilusUtil::count_well_modelled_nas(mol_wrk, predicted_maps);
     std::cout << "Built " << best_na_count << " residues with positive predicted density" << std::endl;
@@ -188,10 +150,11 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
 
         mol_wrk = NucleicAcidTools::flag_chains(mol_wrk);
 
-        mol_wrk = natools.find(xwrk, mol_wrk, nhit / 2, nhit / 2, srchst);
+        clipper::MiniMol nautilus_find = natools.find(xwrk, mol_wrk, nhit / 2, nhit / 2, srchst);
+        mol_wrk = NucleoFind::Find::aggregate_nucleic_nucleic(ml_model, nautilus_find);
         log.log("FIND", mol_wrk, verbose >= 5);
 
-        mol_wrk = run_cycle(nhit, srchst, verbose, natools, seq_wrk, mol_wrk, xwrk, log, false);
+        mol_wrk = run_cycle(verbose, natools, seq_wrk, mol_wrk, xwrk, log, false);
 
         int pred_na_count = NautilusUtil::count_well_modelled_nas(mol_wrk, predicted_maps);
         std::cout << "Cycle "<< cyc+1 << " built " << pred_na_count << " residues with positive predicted density" << std::endl;
@@ -213,7 +176,7 @@ void run(NucleoFind::IO::Input &input, NucleoFind::IO::Output &output, int cycle
     }
 
     // set up residue types
-    mol_wrk = setup_residue_types(xwrk, mol_wrk);
+    mol_wrk = ModelTidy::setup_residue_types(xwrk, mol_wrk);
     log.log("SETRES", mol_wrk, verbose >= 5);
 
     // add true sequence numbers
